Move BitReader class to main utilities namespace
