+++ src/VGAudio/Codecs/CriHca/CriHcaEncryption.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using VGAudio.Utilities;
 
 namespace VGAudio.Codecs.CriHca
 {
